% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionary-function.R
\name{get_val_labels}
\alias{get_val_labels}
\title{Create table of value labels}
\usage{
get_val_labels(data)
}
\arguments{
\item{data}{A data frame or tibble created using `haven::read_sav`}
}
\value{
A data frame with the `variable`, `value` (which has labels), and respective SPSS `value_name`
}
\description{
`get_val_labels` creates a data frame of variable names and labels
}
\details{
This function takes a data frame created using `haven::read_sav` and
returns a data frame with the SPSS labelled values for each variable
}
\keyword{SPSS}
\keyword{haven}
\keyword{sav}
