% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionary-function.R
\name{get_var_labels}
\alias{get_var_labels}
\title{Create table of variable labels}
\usage{
get_var_labels(data)
}
\arguments{
\item{data}{A data frame or tibble created using `haven::read_sav`}
}
\value{
A data frame with the `variable` names and respective SPSS `label`
}
\description{
`get_var_labels` creates a data frame of variable names and labels
}
\details{
This function takes a data frame created using `haven::read_sav` and
returns a data frame with the SPSS variable labels for each variable
}
\keyword{SPSS}
\keyword{haven}
\keyword{sav}
